AI SALES AUTOMATION/my ai automation/ai-sales-automation/requirements.txt
# Core packages
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
pandas==2.0.3
numpy==1.24.4
tqdm==4.66.1

# API clients
tweepy==4.14.0
instagrapi==1.19.4
gspread==5.12.0
oauth2client==4.1.3
stripe==6.2.0
razorpay==1.4.1

# Web automation
selenium==4.11.2
webdriver-manager==3.8.6
beautifulsoup4==4.12.2

# Analytics and tracking
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
